requirements.txt
streamlit==0.76.0
Pillow==8.1.0
shutil
io
base64
datetime